replaces old virtual node implementation with new one for handling multiple paths
